Removed unnecessary min/max finder functions
Today we learned that the >, <, >< scoreboard operations exist, and they have for 10+ years. Not a huge issue, but I wanted to look less dumb by not doing it the hard way.

Also fixed bots going easier on the player after the rival bot finishes. That was opposite of the intended behavior.